requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
couchbase==4.3.4
idna==3.10
panda==0.3.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.6.0
urllib3==2.2.3
